+++ pybamm/models/standard_variables.py
@@ -200,31 +200,31 @@ class StandardVariables:
         self.T_vol_av = pybamm.Variable("Volume-averaged cell temperature")
 
         # SEI variables
-        self.c_inner_av = pybamm.Variable(
-            "X-averaged inner SEI concentration",
+        self.L_inner_av = pybamm.Variable(
+            "X-averaged inner SEI thickness",
             domain="current collector",
         )
-        self.c_inner = pybamm.Variable(
-            "Inner SEI concentration",
+        self.L_inner = pybamm.Variable(
+            "Inner SEI thickness",
             domain=["negative electrode"],
             auxiliary_domains={"secondary": "current collector"},
         )
-        self.c_outer_av = pybamm.Variable(
-            "X-averaged outer SEI concentration",
+        self.L_outer_av = pybamm.Variable(
+            "X-averaged outer SEI thickness",
             domain="current collector",
         )
-        self.c_outer = pybamm.Variable(
-            "Outer SEI concentration",
+        self.L_outer = pybamm.Variable(
+            "Outer SEI thickness",
             domain=["negative electrode"],
             auxiliary_domains={"secondary": "current collector"},
         )
         # For SEI reaction at the li metal/separator interface in a li metal model
-        self.c_inner_interface = pybamm.Variable(
-            "Inner SEI concentration",
+        self.L_inner_interface = pybamm.Variable(
+            "Inner SEI thickness",
             domain=["current collector"],
         )
-        self.c_outer_interface = pybamm.Variable(
-            "Outer SEI concentration",
+        self.L_outer_interface = pybamm.Variable(
+            "Outer SEI thickness",
             domain=["current collector"],
         )