+++ pybamm/models/submodels/interface/sei/constant_sei.py
@@ -31,13 +31,13 @@ class ConstantSEI(BaseModel):
             self.reaction_loc = "full electrode"
 
     def get_fundamental_variables(self):
-        # Constant concentrations
-        c_inner = self.param.L_inner_0
-        c_outer = self.param.L_outer_0
-        variables = self._get_standard_concentration_variables(c_inner, c_outer)
+        # Constant thicknesses
+        L_inner = self.param.L_inner_0
+        L_outer = self.param.L_outer_0
+        variables = self._get_standard_thickness_variables(L_inner, L_outer)
 
-        # Thicknesses (derived from concentrations)
-        variables.update(self._get_standard_thickness_variables(variables))
+        # Concentrations (derived from thicknesses)
+        variables.update(self._get_standard_concentration_variables(variables))
 
         # Reactions
         if self.reaction_loc == "interface":